workable version witout station dependencies --> improve speed, clarity
